+++ admin/ajax.php
@@ -10,7 +10,6 @@
 *               GNU Public License v2 or later
 *   @filesource
 */
-namespace Paypal;
 
 /** Include required glFusion common functions */
 require_once '../../../lib-common.php';
@@ -26,14 +25,13 @@ case 'updatestatus':
     if (!empty($_POST['order_id']) &&
         !empty($_POST['newstatus'])) {
         $showlog = $_POST['showlog'] == 1 ? 1 : 0;
-        USES_paypal_class_Order();
         $log_ts = '';
         $log_user = '';
         $log_msg = '';
         $newstatus = $_POST['newstatus'];
         $order_id = $_POST['order_id'];
         $showlog = $_POST['showlog'] == 1 ? 1 : 0;
-        $ord = new Order($_POST['order_id']);
+        $ord = new Paypal\Order($_POST['order_id']);
         if ($ord->isNew) break;     // non-existant order
         if ($ord->UpdateStatus($newstatus)) {
             $sql = "SELECT * FROM {$_TABLES['paypal.order_log']}
@@ -62,15 +60,13 @@ case 'updatestatus':
 case 'toggle':
     switch ($_POST['component']) {
     case 'product':
-        USES_paypal_class_Product();
-
         switch ($_POST['type']) {
         case 'enabled':
-            $newval = Product::toggleEnabled($_POST['oldval'], $_POST['id']);
+            $newval = Paypal\Product::toggleEnabled($_POST['oldval'], $_POST['id']);
             break;
 
         case 'featured':
-            $newval = Product::toggleFeatured($_POST['oldval'], $_POST['id']);
+            $newval = Paypal\Product::toggleFeatured($_POST['oldval'], $_POST['id']);
             break;
 
          default:
@@ -79,11 +75,9 @@ case 'toggle':
         break;
 
     case 'category':
-        USES_paypal_class_Category();
-
         switch ($_POST['type']) {
         case 'enabled':
-            $newval = Category::toggleEnabled($_REQUEST['oldval'], $_REQUEST['id']);
+            $newval = Paypal\Category::toggleEnabled($_REQUEST['oldval'], $_REQUEST['id']);
             break;
 
          default:
@@ -92,11 +86,9 @@ case 'toggle':
         break;
 
     case 'attribute':
-        USES_paypal_class_Attribute();
-
         switch ($_POST['type']) {
         case 'enabled':
-            $newval = ppAttribute::toggleEnabled($_REQUEST['oldval'], $_REQUEST['id']);
+            $newval = Paypal\Attribute::toggleEnabled($_REQUEST['oldval'], $_REQUEST['id']);
             break;
 
          default:
@@ -106,19 +98,17 @@ case 'toggle':
        break;
 
     case 'gateway':
-        USES_paypal_gateway_base();
-
         switch ($_POST['type']) {
         case 'enabled':
-            $newval = PaymentGw::toggleEnabled($_REQUEST['oldval'], $_REQUEST['id']);
+            $newval = Paypal\Gateway::toggleEnabled($_REQUEST['oldval'], $_REQUEST['id']);
             break;
 
         case 'buy_now':
-            $newval = PaymentGw::toggleBuyNow($_REQUEST['oldval'], $_REQUEST['id']);
+            $newval = Paypal\Gateway::toggleBuyNow($_REQUEST['oldval'], $_REQUEST['id']);
             break;
 
         case 'donation':
-            $newval = PaymentGw::toggleDonation($_REQUEST['oldval'], $_REQUEST['id']);
+            $newval = Paypal\Gateway::toggleDonation($_REQUEST['oldval'], $_REQUEST['id']);
             break;
 
         default:
@@ -127,11 +117,10 @@ case 'toggle':
         break;
 
     case 'workflow':
-        USES_paypal_class_Workflow();
         $field = $_POST['type'];
         switch ($field) {
         case 'enabled':
-            $newval = Workflow::Toggle($_REQUEST['id'], $field, $_REQUEST['oldval']);
+            $newval = Paypal\Workflow::Toggle($_REQUEST['id'], $field, $_REQUEST['oldval']);
             break;
 
         default:
@@ -140,12 +129,11 @@ case 'toggle':
         break;
 
     case 'orderstatus':
-        USES_paypal_class_OrderStatus();
         $field = $_POST['type'];
         switch ($field) {
         case 'enabled':
         case 'notify_buyer':
-            $newval = OrderStatus::Toggle($_REQUEST['id'], $field, $_REQUEST['oldval']);
+            $newval = Paypal\OrderStatus::Toggle($_REQUEST['id'], $field, $_REQUEST['oldval']);
             break;
 
         default: