+++ language/german.php
@@ -287,6 +287,7 @@ $LANG_PP = array (
 'msg_item_added' => 'Item has been added to your cart',
 'all' => 'All',
 'print' => 'Print',
+'resetbuttons' => 'Clear Encrypted Button Cache',
 );
 if ($_PP_CONF['ena_ratings']) {
     $LANG_PP['list_sort_options']['top_rated'] = 'Top Rated';