+++ public_html/index.php
@@ -15,7 +15,6 @@
 *               GNU Public License v2 or later
 *   @filesource
 */
-namespace Paypal;
 
 /** Require core glFusion code */
 require_once '../lib-common.php';
@@ -94,8 +93,6 @@ case 'checkout':
         $ppGCart->setInstructions($_POST['order_instr']);
     }
     if ($_PP_CONF['anon_buy'] == 1 || !COM_isAnonUser()) {
-        USES_paypal_class_Workflow();
-        USES_paypal_class_UserInfo();
         // Start with the first view.
         //$view = Workflow::getNextView();
         $view = 'checkoutcart';
@@ -105,7 +102,7 @@ case 'checkout':
             switch ($wf_name) {
             case 'billto':
             case 'shipto':
-                if (!(UserInfo::isValidAddress($ppGCart->getAddress($wf_name)) == '')) {
+                if (!(Paypal\UserInfo::isValidAddress($ppGCart->getAddress($wf_name)) == '')) {
                     //$view = Workflow::getNextView($wf_name);
                     $view = $wf_name;
                     break 2;    // exit switch and foreach
@@ -122,9 +119,7 @@ case 'checkout':
 case 'savebillto':
 case 'saveshipto':
     $addr_type = substr($action, 4);   // get 'billto' or 'shipto'
-    USES_paypal_class_UserInfo();
-    USES_paypal_class_Workflow();
-    $status = UserInfo::isValidAddress($_POST);
+    $status = Paypal\UserInfo::isValidAddress($_POST);
     if ($status != '') {
         $content .= PAYPAL_errMsg($status, $LANG_PP['invalid_form']);
         $view = $addr_type;
@@ -144,13 +139,12 @@ case 'saveshipto':
             $_POST['useaddress'] = $addr_id[0];
         }
     }
-    $view = Workflow::getNextView($addr_type);
+    $view = Paypal\Workflow::getNextView($addr_type);
     $ppGCart->setAddress($_POST, $addr_type);
     break;
 
 case 'addcartitem':
 case 'addcartitem_x':   // using the image submit button, such as Paypal's
-    USES_paypal_class_Cart();
     $view = 'productlist';
     if (isset($_POST['_unique']) && $_POST['_unique'] &&
             $ppGCart->Contains($_POST['item_number']) !== false) {
@@ -176,8 +170,7 @@ case 'delcartitem':
     $ppGCart->Remove($_GET['id']);
     /*if (isset($_SESSION[PP_CART_VAR]['order_id']) &&
         !empty($_SESSION[PP_CART_VAR]['order_id'])) {
-        //USES_paypal_class_Order();
-        //Order::delCartItem($_GET['id'], $_SESSION[PP_CART_VAR]['order_id']);
+        //Paypal\Order::delCartItem($_GET['id'], $_SESSION[PP_CART_VAR]['order_id']);
         $ppGCart->Remove($_GET['id']);
     }*/
     $view = 'cart';
@@ -195,31 +188,28 @@ case 'emptycart':
     break;
 
 case 'thanks':
-    if (!empty($actionval) && USES_paypal_gateway($actionval)) {
-        $classname = '\\Paypal\\' . $actionval;
-        $gw = new $classname();
-        $tVars = $gw->thanksVars();
-        if (!empty($tVars)) {
-            $T = new \Template($_CONF['path'] . 'plugins/paypal/templates');
-            $T ->set_file(array('msg'   => 'thanks_for_order.thtml'));
-            $T->set_var(array(
-                'site_name'     => $_CONF['site_name'],
-                'payment_date'  => $tVars['payment_date'],
-                'currency'      => $tVars['currency'],
-                'mc_gross'      => $tVars['payment_amount'],
-                'gateway_url'   => $tVars['gateway_url'],
-                'gateway_name'  => $tVars['gateway_name'],
-                'payment_status' => $tVars['payment_status'],
-                'completed' => $tVars['_status'] == 'completed' ? 'true' : '',
-            ) );
-            $message = $T->parse('output', 'msg');
-        } else {
-            // Allow for no thanksVars function
-            $message = $LANG_PP['thanks_title'];
+    // Allow for no thanksVars function
+    $message = $LANG_PP['thanks_title'];
+    if (!empty($actionval)) {
+        $gw = Paypal\Gateway($actionval);
+        if ($gw !== NULL) {
+            $tVars = $gw->thanksVars();
+            if (!empty($tVars)) {
+                $T = new \Template($_CONF['path'] . 'plugins/paypal/templates');
+                $T ->set_file(array('msg'   => 'thanks_for_order.thtml'));
+                $T->set_var(array(
+                    'site_name'     => $_CONF['site_name'],
+                    'payment_date'  => $tVars['payment_date'],
+                    'currency'      => $tVars['currency'],
+                    'mc_gross'      => $tVars['payment_amount'],
+                    'gateway_url'   => $tVars['gateway_url'],
+                    'gateway_name'  => $tVars['gateway_name'],
+                    'payment_status' => $tVars['payment_status'],
+                    'completed' => $tVars['_status'] == 'completed' ? 'true' : '',
+                ) );
+                $message = $T->parse('output', 'msg');
+            }
         }
-    } else {
-        // Allow for missing or unknown payment gateway name
-        $message = $LANG_PP['thanks_title'];
     }
     $content .= COM_showMessageText($message, $LANG_PP['thanks_title'], true, 'success');
     $view = 'productlist';
@@ -253,10 +243,8 @@ case 'processorder':
     // Process the order, similar to what an IPN would normally do.
     // This is for internal, manual processes like C.O.D. or Prepayment orders
     $gw_name = isset($_POST['gateway']) ? $_POST['gateway'] : '';
-    $status = USES_paypal_gateway($gw_name);
-    if ($status) {
-        $classname = '\\Paypal\\' . $gw_name;
-        $gw = new $classname;
+    $gw = Paypal\Gateway::getInstance($gw_name);
+    if ($gw !== NULL) {
         $output = $gw->handlePurchase($_POST);
         if (!empty($output)) {
             $content .= $output;
@@ -265,7 +253,7 @@ case 'processorder':
         }
         $view = 'thanks';
         $ppGCart->Clear(false);
-        if (USES_paypal_gateway($actionval)) {
+        /*if (USES_paypal_gateway($actionval)) {
             $gw = new $actionval;
             $tVars = $gw->thanksVars();
             if (!empty($tVars)) {
@@ -287,8 +275,8 @@ case 'processorder':
         } else {
             // Allow for missing or unknown payment gateway name
             $message = $LANG_PP['thanks_title'];
-        }
-        $content .= COM_showMessageText($message, $LANG_PP['thanks_title'], true);
+        }*/
+        //$content .= COM_showMessageText($message, $LANG_PP['thanks_title'], true);
     }
     $view = 'productlist';
     break;
@@ -310,16 +298,14 @@ case 'history':
 case 'billto':
 case 'shipto':
     if (COM_isAnonUser()) COM_404();
-    USES_paypal_class_UserInfo();
-    $U = new UserInfo();
+    $U = new Paypal\UserInfo();
     $A = isset($_POST['address1']) ? $_POST : $ppGCart->getAddress($view);
     $content .= $U->AddressForm($view, $A);
     break;
 
 case 'order':
     if (COM_isAnonUser()) COM_404();
-    USES_paypal_class_Order();
-    $order = new Order($actionval);
+    $order = new Paypal\Order($actionval);
     if ($order->canView()) {
         $content .= $order->View(true);
     } else {
@@ -328,8 +314,7 @@ case 'order':
     break;
 
 case 'printorder':
-    USES_paypal_class_Order();
-    $order = new Order($actionval);
+    $order = new Paypal\Order($actionval);
     if ($order->canView()) {
         echo $order->View(true, 'print');
         exit;
@@ -341,15 +326,14 @@ case 'printorder':
 case 'vieworder':
     if (COM_isAnonUser()) COM_404();
     //$_SESSION[PP_CART_VAR]['prevpage'] = $view;
-    Cart::setSession('prevpage', $view);
+    Paypal\Cart::setSession('prevpage', $view);
     $content .= $ppGCart->View(true);
     $page_title = $LANG_PP['view_order'];
     break;
 
 case 'detail':
     // deprecated, should be displayed via detail.php
-    USES_paypal_class_Product();
-    $P = new Product($id);
+    $P = new Paypal\Product($id);
     $content .= $P->Detail();
     $menu_opt = $LANG_PP['product_list'];
     $page_title = $LANG_PP['product_detail'];
@@ -370,9 +354,9 @@ case 'viewcart':
 case 'checkoutcart':
     // Need to create an order or save the cart, so IPN class
     // can access the data. For now, use the cart.
-    /*USES_paypal_class_Order();
+    /*
     if (empty($_SESSION[PP_CART_VAR]['invoice'])) {
-        $Ord = new Order();
+        $Ord = new Paypal\Order();
         $Ord->CreateFromCart($ppGCart);
     } else {
         $Ord = new Order($_SESSION[PP_CART_VAR]['invoice']);
@@ -384,9 +368,9 @@ case 'checkoutcart':
 case 'productlist':
 default:
     if (isset($_REQUEST['category'])) {
-        $content .= ProductList($_REQUEST['category']);
+        $content .= Paypal\ProductList($_REQUEST['category']);
     } else {
-        $content .= ProductList();
+        $content .= Paypal\ProductList();
     }
     $menu_opt = $LANG_PP['product_list'];
     $page_title = $LANG_PP['main_title'];
@@ -397,8 +381,8 @@ case 'none':
     break;
 }
 
-$display = siteHeader();
-$T = new \Template(PAYPAL_PI_PATH . '/templates');
+$display = Paypal\siteHeader();
+$T = new Template(PAYPAL_PI_PATH . '/templates');
 $T->set_file('title', 'paypal_title.thtml');
 $T->set_var('title', $page_title);
 $display .= $T->parse('', 'title');
@@ -410,7 +394,7 @@ if (!empty($msg)) {
 }
 $display .= LGLIB_showAllMessages();
 $display .= $content;
-$display .= siteFooter();
+$display .= Paypal\siteFooter();
 echo $display;
 
 ?>